dumpview: allow opening multiple swfs, context menu: expand all and close menu items

+++ src/com/jpexs/decompiler/flash/dumpview/DumpInfo.java
@@ -42,7 +42,7 @@ public class DumpInfo {
 
     public List<DumpInfo> childInfos = new ArrayList<>();
 
-    public DumpInfo(String name, String type, Object value, long startByte, int lengthBytes) {
+    public DumpInfo(String name, String type, Object value, long startByte, long lengthBytes) {
 
         this.name = name;
         this.type = type;
